SmartMeasure: AR-Stats-Panel und Tracking-Reasons lokalisieren (DE/EN/ES/FR/IT/PT)
- 14 neue RESX-Keys (ArStat*, ArReticleHoldStill, ArBadgeReady, ArTracking*) in
  AppStrings.resx + .en/.es/.fr/.it/.pt + Designer-Properties.
- Neuer ArOverlayLabels-Record mit Session-Snapshot (Sprachwechsel mid-AR-Session
  gibts nicht — Activity laeuft modal-fullscreen) + German-Defaults-Property.
- ArCaptureActivity.LoadLocalizedLabels liest die Strings beim OnCreate aus
  AppStrings (CurrentUICulture ist da bereits gesetzt), reicht den Snapshot in
  jedem BuildOverlayState durch.
- ArPointOverlayView nutzt _state.Labels statt hardcoded DE-Strings in
  DrawStatsPanel, DrawLiveFooter, DrawReticle (Sampling-Progress), DrawReadinessBadge.
- TrackingFailureReason wird aus dem Labels-Snapshot lokalisiert.

+++ src/Apps/FinanzRechner/FinanzRechner.Shared/Models/Account.cs
@@ -26,7 +26,7 @@ public class Account
     public AccountType Type { get; set; } = AccountType.Checking;
 
     /// <summary>Anfangssaldo beim Erstellen des Kontos.</summary>
-    public double InitialBalance { get; set; }
+    public decimal InitialBalance { get; set; }
 
     /// <summary>Emoji-Icon für die Darstellung.</summary>
     public string Icon { get; set; } = "\U0001F3E6"; // Bankgebäude
@@ -51,10 +51,10 @@ public class Account
 /// </summary>
 public record AccountBalance(
     Account Account,
-    double CurrentBalance,
-    double MonthlyIncome,
-    double MonthlyExpenses,
-    double MonthlyBalance)
+    decimal CurrentBalance,
+    decimal MonthlyIncome,
+    decimal MonthlyExpenses,
+    decimal MonthlyBalance)
 {
     public string BalanceDisplay => CurrencyHelper.Format(CurrentBalance);
     public string MonthlyIncomeDisplay => CurrencyHelper.FormatSigned(MonthlyIncome);
