+++ src/Apps/FinanzRechner/FinanzRechner.Shared/Services/NotificationService.cs
@@ -2,7 +2,7 @@ namespace FinanzRechner.Services;
 
 public sealed class NotificationService : INotificationService
 {
-    public Task SendBudgetAlertAsync(string categoryName, double percentageUsed, double spent, double limit)
+    public Task SendBudgetAlertAsync(string categoryName, decimal percentageUsed, decimal spent, decimal limit)
     {
         // Desktop: Keine Aktion, kann mit plattformspezifischen Benachrichtigungen erweitert werden
         return Task.CompletedTask;
